optimising looping through df.
